chore: bump basedpyright version
New version also adds a new option. This option modifies the type
system. Currently there is only one rule that actually utilizes this
change. But more will be added in the future.

As a note; you should not enable this rule if you are a library. But
seeing as this is not a library, it makes sense to enable.