style: explicit boolean evaluation return type
This is solving some errors that `ty` is reporting. Entity could have a
subclass that didn't evaluate to true every time. So we explicitly state
that if the object exists, it is a truthy object. This resolves type
errors around (entity and Comment).

The `message_filter` could return `None`. `None and x` returns `None`.
None is not a valid boolean. Instead I convert it explicitly to a
boolean.